System関係更新

+++ ShadowStepGame/MapSystemComponent.h
@@ -73,6 +73,6 @@ public:
     // ===================================================================
     float GetMapSizeHeight() const { return (float)m_MapHeight * m_SizePiece; }
     float GetMapSizeWidth() const { return (float)m_MapWidth * m_SizePiece; }
-    float GetMapHeight() const { return (float)m_MapHeight; }
-    float GetMapWidth() const { return (float)m_MapWidth; }
+    int GetMapHeight() const { return m_MapHeight; }
+    int GetMapWidth() const { return m_MapWidth; }
 };
